+++ branch-alpha/jslint.mjs
@@ -165,7 +165,7 @@ let jslint_charset_ascii = (
     + "@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_"
     + "`abcdefghijklmnopqrstuvwxyz{|}~\u007f"
 );
-let jslint_edition = "v2023.1.29";
+let jslint_edition = "v2022.4.1-beta";
 let jslint_export;                      // The jslint object to be exported.
 let jslint_fudge = 1;                   // Fudge starting line and starting
                                         // ... column to 1.
@@ -11236,25 +11236,27 @@ function sentinel() {}
             }
         }));
         exitCode = await new Promise(function (resolve) {
-            moduleChildProcess.spawn((
-                processArgv[0] === "npm"
+            moduleChildProcess.spawn(
+                (
+                    processArgv[0] === "npm"
 
 // If win32 environment, then replace program npm with npm.cmd.
 // Coverage-hack - Ugly-hack to get test-coverage under both win32 and linux.
 
-                ? process.platform.replace("win32", "npm.cmd").replace(
-                    process.platform,
-                    "npm"
-                )
-                : processArgv[0]
-            ), processArgv.slice(1), {
-                env: Object.assign({}, process.env, {
-                    NODE_V8_COVERAGE: coverageDir
-                }),
-                stdio: [
-                    "ignore", 1, 2
-                ]
-            }).on("exit", resolve);
+                    ? process.platform.replace("win32", "npm.cmd").replace(
+                        process.platform,
+                        "npm"
+                    )
+                    : processArgv[0]
+                ),
+                processArgv.slice(1),
+                {
+                    env: Object.assign({}, process.env, {
+                        NODE_V8_COVERAGE: coverageDir
+                    }),
+                    stdio: ["ignore", 1, 2]
+                }
+            ).on("exit", resolve);
         });
     }