+++ branch-alpha/CHANGELOG.md
@@ -9,6 +9,11 @@
 - jslint - try to improve parser to be able to parse jquery.js without stopping.
 - jslint - unify analysis of variable-assignment/function-parameters into one function
 
+# v2022.4.1-beta
+- ci - housekeeping of ci-file jslint_ci.sh
+- ci - update shell-function shGitCommitPushOrSquash() to be able to save to specified filename
+- ci - remove little-used shell-function shCiBranchPromote()
+
 # v2023.1.29
 - ci - in windows-ci-env, alias node=node.exe instead of using winpty for pipes
 - ci - bugfix - fix ci-shell-function shGithubFileUpload unable to upload new asset